requirements.txt
python-dotenv
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Deprecated==1.2.14
frozenlist==1.4.1
github @ git+https://github.com/VarMonke/Github-Api-Wrapper@a1ef2e96f0f5df77340de69e9fe66c10410284cb
idna==3.7
multidict==6.0.5
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.32.3
typing_extensions==4.12.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
